requirements.txt
fastapi>=0.95.1
Mako>=1.2.4
MarkupSafe>=2.1.0
mongoengine>=0.27.0
mysql-connector>=2.2.9
mysql-connector-python>=8.0.33
mysql-connector-python-rf>=2.2.2
pydantic>=1.10.7
pymongo>=4.3.3
SQLAlchemy>=2.0.10
uvicorn>=0.15.0
python-dotenv==1.0.0
virtualenv>=20.17.1
